ast refactor Element List v2
