--- server/src/ast/v2/Condition/ConditionArgument.ts
@@ -1,8 +1,0 @@
-import Condition from "betonquest-utils/betonquest/Condition";
-
-import { ConditionArgumentType } from "../../node";
-import { ElementArgument } from "../Element/ElementArgument";
-
-export abstract class ConditionArgument extends ElementArgument<Condition> {
-  type: ConditionArgumentType = "ConditionArgument";
-}
