--- server/src/ast/v2/Element/ElementArgumentMandatory.ts
@@ -1,10 +1,0 @@
-import ListElement from "betonquest-utils/betonquest/ListElement";
-
-import { ElementArgument } from "./ElementArgument";
-
-export abstract class ElementArgumentMandatory<LE extends ListElement> extends ElementArgument<LE> {
-  getSemanticTokens() {
-    // TODO
-    return [];
-  }
-}
