--- server/src/ast/v2/Element/ElementArgumentOptional.ts
@@ -1,10 +1,0 @@
-import ListElement from "betonquest-utils/betonquest/ListElement";
-
-import { ElementArgument } from "./ElementArgument";
-
-export abstract class ElementArgumentOptional<LE extends ListElement> extends ElementArgument<LE> {
-  getSemanticTokens() {
-    // TODO
-    return [];
-  }
-}
