+++ server/src/ast/v2/Event/EventKey.ts
@@ -1,11 +1,30 @@
-import Event from "betonquest-utils/betonquest/Event";
 
-import { EventKeyType } from "../../node";
-import { ElementKey } from "../Element/ElementKey";
+import { Scalar } from "yaml";
+
 import { SemanticToken, SemanticTokenType } from "../../../service/semanticTokens";
+import { HoverInfo } from "../../../utils/hover";
+import { EventKeyType } from "../../node";
+import { AbstractNodeV2 } from "../../v2";
+import { EventEntry } from "./EventEntry";
 
-export class EventKey extends ElementKey<Event> {
-  type: EventKeyType = "EventKey";
+export class EventKey extends AbstractNodeV2<EventKeyType> {
+  readonly type: EventKeyType = "EventKey";
+  readonly offsetStart: number;
+  readonly offsetEnd: number;
+  readonly parent: EventEntry;
+
+  private value: string;
+  private comment?: string;
+
+  constructor(key: Scalar<string>, parent: EventEntry) {
+    super();
+    this.offsetStart = key.range![0];
+    this.offsetEnd = key.range![1];
+    this.parent = parent;
+
+    this.value = key.value;
+    this.comment = key.commentBefore?.split(/\n\n+/).slice(-1)[0] ?? undefined;
+  }
 
   getSemanticTokens(): SemanticToken[] {
     if (this.offsetStart === undefined || this.offsetEnd === undefined) {
@@ -17,4 +36,20 @@ export class EventKey extends ElementKey<Event> {
       tokenType: SemanticTokenType.EventID
     }];
   };
+  getHoverInfo(offset?: number): HoverInfo[] {
+    if (!offset || this.offsetStart <= offset && this.offsetEnd >= offset) {
+      const hoverInfo: HoverInfo[] = [{
+        content: "(full path: " + this.value + ")",
+        offset: [this.offsetStart, this.offsetEnd]
+      }];
+      if (this.comment) {
+        hoverInfo.unshift({
+          content: this.comment,
+          offset: [this.offsetStart!, this.offsetEnd!]
+        });
+      }
+      return hoverInfo;
+    }
+    return [];
+  }
 }
