+++ server/src/ast/v2/Objective/ObjectiveArgumentOptional.ts
@@ -1,10 +1,32 @@
-import Objective from "betonquest-utils/betonquest/Objective";
 
-import { ObjectiveArgumentType } from "../../node";
-import { ElementArgumentOptional } from "../Element/ElementArgumentOptional";
+import { ArgumentsPatternOptional } from "betonquest-utils/betonquest/Arguments";
+import { ObjectiveArgumentOptionalType } from "../../node";
+import { AbstractNodeV2 } from "../../v2";
+import { ObjectiveArguments } from "./ObjectiveArguments";
 
-export class ObjectiveArgumentOptional extends ElementArgumentOptional<Objective> {
-  type: ObjectiveArgumentType = 'ObjectiveArgument';
+export class ObjectiveArgumentOptional extends AbstractNodeV2<ObjectiveArgumentOptionalType> {
+  readonly type: ObjectiveArgumentOptionalType = 'ObjectiveArgumentOptional'; // TODO remove Mandatory / Optional
+  offsetStart?: number;
+  offsetEnd?: number;
+  readonly parent: ObjectiveArguments;
+
+  argumentStr: string;
+
+  constructor(argumentStr: string,
+    range: [number?, number?],
+    // isMandatory: boolean,
+    pattern: ArgumentsPatternOptional,
+    parent: ObjectiveArguments,
+  ) {
+    super();
+    this.parent = parent;
+
+    this.offsetStart = range[0];
+    this.offsetEnd = range[1];
+
+    // Parse argumentStr
+    this.argumentStr = argumentStr;
+
+    // Check format
+  }
 }
-
-// new ObjectiveArgumentOptional().parent;
